Finishes first prototype of rusqlv2!
We must have library, sorry. It does make it easier to link against it (MySql and Boost are linked automatically).

Most of your actions can be performed directly on the database, no need to create a statement or that kind of stuff.
Example:
	database->query("select * from table where id=?", id).get_uint64("column_name");
Bam you just got the first row's column_name as a uint64_t or a rusql::NoResults exception. Congrats.

I have plenty of examples ready to show, they will come!